requirements.txt
alembic==0.8.9
bcrypt==3.1.1
cffi==1.9.1
click==6.6
Flask==0.11.1
Flask-Migrate==2.0.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
Pillow==4.0.0
pycparser==2.17
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.11
WTForms==2.1
